[bugfix] start

+++ start.py
@@ -44,6 +44,7 @@ def git_update():
     subprocess.check_output([git_cmd, '-C', str(pro_dir.resolve()), 'checkout', '.'])
     subprocess.check_output([git_cmd, '-C', str(pro_dir.resolve()), 'pull', 'origin', 'master'])
     for p in history_save_dir.glob('**/*'):
+        if p.is_dir(): continue
         cmp_p = tutorial_dir / pathlib.Path(*p.relative_to(history_save_dir).parts[1:])
         if cmp_p.exists() and filecmp.cmp(p, cmp_p):
             shutil.rmtree(p)
